Django/07_Django StaticMedia/Django StaticMedia/requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.9.3
colorama==0.4.4
debugpy==1.6.2
decorator==5.1.1
Django==3.2.13
django-appconf==1.0.5
django-bootstrap-v5==1.0.2
django-extensions==3.1.5
django-imagekit==4.1.0
django-seed==0.2.2
entrypoints==0.4
executing==0.8.3
Faker==8.1.1
importlib-metadata==2.1.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
prompt-toolkit==3.0.29
psutil==5.9.1
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.1
pytz==2022.1
pywin32==304
pyzmq==23.2.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
stack-data==0.2.0
text-unidecode==1.3
tornado==6.2
traitlets==5.2.2.post1
wcwidth==0.2.5
zipp==3.4.1
